+++ app/controllers/kitten/utilities_controller.rb
@@ -1,0 +1,8 @@
+module Kitten
+  class UtilitiesController < Kitten::ApplicationController
+    def index
+      render template: "layouts/kitten/utilities",
+             layout: "kitten/application"
+    end
+  end
+end